refactor(router): split worker selector modules
