+++ lib/kv-router/src/scheduling/selector/default.rs
@@ -2,35 +2,22 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use std::borrow::Borrow;
-use std::cell::RefCell;
 use std::collections::HashMap;
-use std::ops::BitOr;
 #[cfg(any(test, feature = "bench"))]
 use std::sync::Arc;
 
 use parking_lot::Mutex;
-#[cfg(test)]
-use rustc_hash::FxHashMap;
 
-use super::config::KvRouterConfig;
-use super::filter::{RoutingEligibility, WorkerEligibilityError};
-use super::types::{KvSchedulerError, SchedulingRequest, WorkerSelectionPolicyError};
+use super::policy::WorkerSelectionPolicyStateRef;
+use super::{
+    LogitWeights, ScoredWorkerCandidate, WorkerCandidate, WorkerInputView, WorkerInputs,
+    WorkerPicker, WorkerScorer, WorkerSelectionContext, WorkerSelectionInput, WorkerSelector,
+    select_worker_with_policy,
+};
 use crate::protocols::{WorkerConfigLike, WorkerId, WorkerSelectionResult, WorkerWithDpRank};
-
-/// Low-level worker-selection contract used by the scheduler.
-///
-/// Generic over `C` so that the scheduling layer does not depend on a concrete config type.
-/// External policies should compose [`WorkerScorer`] and [`WorkerPicker`] implementations with
-/// [`WorkerSelectionPolicy`] instead of implementing this trait directly.
-pub trait WorkerSelector<C: WorkerConfigLike> {
-    fn select_worker(
-        &self,
-        workers: &HashMap<WorkerId, C>,
-        request: &SchedulingRequest,
-        eligibility: RoutingEligibility<'_>,
-        block_size: u32,
-    ) -> Result<WorkerSelectionResult, KvSchedulerError>;
-}
+use crate::scheduling::config::KvRouterConfig;
+use crate::scheduling::filter::RoutingEligibility;
+use crate::scheduling::types::{KvSchedulerError, SchedulingRequest, WorkerSelectionPolicyError};
 
 #[cfg(any(test, feature = "bench"))]
 fn softmax_sample_entries<T: Copy>(
@@ -109,41 +96,9 @@ pub struct DefaultWorkerSelector {
     picker: DefaultWorkerPicker,
 }
 
-#[derive(Debug, Clone, Copy)]
-struct LogitWeights {
-    overlap_score_credit: f64,
-    overlap_score_credit_decay: f64,
-    prefill_load_scale: f64,
-    shared_cache_multiplier: f64,
-}
-
-struct WorkerSelectionInput<'a> {
-    request: &'a SchedulingRequest,
-    has_tier_overlap_blocks: bool,
-    context: WorkerSelectionContext<'a>,
-}
-
-pub struct WorkerSelectionContext<'a> {
-    request_id: &'a str,
-    request_blocks: u64,
-    block_size: u32,
-    track_prefill_tokens: bool,
-    weights: LogitWeights,
-    min_active_prefill_tokens: usize,
-    router_temperature_override: Option<f64>,
-}
-
-pub struct WorkerCandidate {
-    worker: WorkerWithDpRank,
-    inputs: WorkerInputs,
-    cache: WorkerCacheInput,
-    load: WorkerLoadInput,
-    routing: WorkerRoutingInput,
-}
-
 pub struct DefaultWorkerScorer<C = KvRouterConfig> {
-    kv_router_config: C,
-    worker_type: &'static str,
+    pub(super) kv_router_config: C,
+    pub(super) worker_type: &'static str,
 }
 
 pub struct DefaultWorkerPicker {
@@ -158,285 +113,6 @@ pub struct DefaultWorkerPicker {
 struct DefaultSoftmaxScratch {
     entries: Vec<(WorkerWithDpRank, f64)>,
     probabilities: Vec<f64>,
-}
-
-#[derive(Clone, Copy)]
-pub struct ScoredWorkerCandidate {
-    worker: WorkerWithDpRank,
-    cost: f64,
-}
-
-/// Optional worker-signal groups requested by scorers and pickers.
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
-pub struct WorkerInputs(u8);
-
-impl WorkerInputs {
-    pub const NONE: Self = Self(0);
-    pub const CACHE: Self = Self(1 << 0);
-    pub const LOAD: Self = Self(1 << 1);
-    pub const ROUTING: Self = Self(1 << 2);
-    pub const ALL: Self = Self(Self::CACHE.0 | Self::LOAD.0 | Self::ROUTING.0);
-
-    pub const fn contains(self, other: Self) -> bool {
-        self.0 & other.0 == other.0
-    }
-}
-
-impl BitOr for WorkerInputs {
-    type Output = Self;
-
-    fn bitor(self, rhs: Self) -> Self::Output {
-        Self(self.0 | rhs.0)
-    }
-}
-
-#[derive(Clone, Copy, Default)]
-pub struct WorkerCacheInput {
-    effective_overlap_blocks: f64,
-    device_overlap_blocks: f64,
-    host_overlap_blocks: f64,
-    disk_overlap_blocks: f64,
-    shared_beyond_device_blocks: u32,
-}
-
-#[derive(Clone, Copy, Default)]
-pub struct WorkerLoadInput {
-    raw_prefill_blocks: f64,
-    active_prefill_tokens: usize,
-    decode_cost_blocks: f64,
-    active_requests: usize,
-}
-
-#[derive(Clone, Copy, Default)]
-pub struct WorkerRoutingInput {
-    preferred_taint_multiplier: Option<f64>,
-}
-
-/// Borrowed, index-aligned view of one custom picker's requested worker inputs.
-#[derive(Clone, Copy)]
-pub struct WorkerInputView<'a> {
-    candidates: &'a [ScoredWorkerCandidate],
-    cache: Option<&'a [WorkerCacheInput]>,
-    load: Option<&'a [WorkerLoadInput]>,
-    routing: Option<&'a [WorkerRoutingInput]>,
-}
-
-pub trait WorkerScorer: Send {
-    /// Declare the worker-signal groups needed by this scorer.
-    fn required_worker_inputs(&self) -> WorkerInputs {
-        WorkerInputs::NONE
-    }
-
-    /// Return one finite, lower-is-better cost contribution for an eligible worker row.
-    fn score(
-        &mut self,
-        context: &WorkerSelectionContext<'_>,
-        candidate: &WorkerCandidate,
-    ) -> Result<f64, WorkerSelectionPolicyError>;
-}
-
-pub trait WorkerPicker: Send {
-    /// Declare the optional worker-signal columns needed by this picker.
-    fn required_worker_inputs(&self) -> WorkerInputs {
-        WorkerInputs::NONE
-    }
-
-    /// Return one row index from the host-owned eligible candidate table.
-    fn pick(
-        &mut self,
-        context: &WorkerSelectionContext<'_>,
-        input: WorkerInputView<'_>,
-    ) -> Result<usize, WorkerSelectionPolicyError>;
-}
-
-impl WorkerSelectionContext<'_> {
-    pub fn request_id(&self) -> &str {
-        self.request_id
-    }
-
-    pub fn request_blocks(&self) -> u64 {
-        self.request_blocks
-    }
-
-    pub fn block_size(&self) -> u32 {
-        self.block_size
-    }
-
-    pub fn tracks_prefill_tokens(&self) -> bool {
-        self.track_prefill_tokens
-    }
-
-    pub fn router_temperature_override(&self) -> Option<f64> {
-        self.router_temperature_override
-    }
-}
-
-impl WorkerCandidate {
-    pub fn worker(&self) -> WorkerWithDpRank {
-        self.worker
-    }
-
-    pub fn cache(&self) -> Option<&WorkerCacheInput> {
-        self.inputs
-            .contains(WorkerInputs::CACHE)
-            .then_some(&self.cache)
-    }
-
-    pub fn load(&self) -> Option<&WorkerLoadInput> {
-        self.inputs
-            .contains(WorkerInputs::LOAD)
-            .then_some(&self.load)
-    }
-
-    pub fn routing(&self) -> Option<&WorkerRoutingInput> {
-        self.inputs
-            .contains(WorkerInputs::ROUTING)
-            .then_some(&self.routing)
-    }
-}
-
-impl ScoredWorkerCandidate {
-    pub fn worker(&self) -> WorkerWithDpRank {
-        self.worker
-    }
-
-    pub fn cost(&self) -> f64 {
-        self.cost
-    }
-}
-
-impl WorkerCacheInput {
-    pub fn effective_overlap_blocks(&self) -> f64 {
-        self.effective_overlap_blocks
-    }
-
-    pub fn device_overlap_blocks(&self) -> f64 {
-        self.device_overlap_blocks
-    }
-
-    pub fn host_overlap_blocks(&self) -> f64 {
-        self.host_overlap_blocks
-    }
-
-    pub fn disk_overlap_blocks(&self) -> f64 {
-        self.disk_overlap_blocks
-    }
-
-    pub fn shared_beyond_device_blocks(&self) -> u32 {
-        self.shared_beyond_device_blocks
-    }
-}
-
-impl WorkerLoadInput {
-    pub fn raw_prefill_blocks(&self) -> f64 {
-        self.raw_prefill_blocks
-    }
-
-    pub fn active_prefill_tokens(&self) -> usize {
-        self.active_prefill_tokens
-    }
-
-    pub fn decode_cost_blocks(&self) -> f64 {
-        self.decode_cost_blocks
-    }
-
-    pub fn active_requests(&self) -> usize {
-        self.active_requests
-    }
-}
-
-impl WorkerRoutingInput {
-    pub fn preferred_taint_multiplier(&self) -> Option<f64> {
-        self.preferred_taint_multiplier
-    }
-}
-
-impl<'a> WorkerInputView<'a> {
-    pub fn candidates(self) -> &'a [ScoredWorkerCandidate] {
-        self.candidates
-    }
-
-    pub fn cache(self) -> Option<&'a [WorkerCacheInput]> {
-        self.cache
-    }
-
-    pub fn load(self) -> Option<&'a [WorkerLoadInput]> {
-        self.load
-    }
-
-    pub fn routing(self) -> Option<&'a [WorkerRoutingInput]> {
-        self.routing
-    }
-}
-
-#[cfg_attr(not(feature = "standalone-selection"), allow(dead_code))]
-enum WorkerSelectionPolicyState {
-    Default(DefaultWorkerPicker),
-    Custom(RefCell<CustomWorkerSelectionState>),
-}
-
-enum WorkerSelectionPolicyStateRef<'a> {
-    Default(&'a DefaultWorkerPicker),
-    Custom(&'a RefCell<CustomWorkerSelectionState>),
-}
-
-struct CustomWorkerSelectionState {
-    scorers: Vec<Box<dyn WorkerScorer>>,
-    picker: Box<dyn WorkerPicker>,
-    worker_inputs: WorkerInputs,
-    picker_inputs: WorkerInputs,
-    candidates: Vec<ScoredWorkerCandidate>,
-    cache_inputs: Vec<WorkerCacheInput>,
-    load_inputs: Vec<WorkerLoadInput>,
-    routing_inputs: Vec<WorkerRoutingInput>,
-}
-
-/// Native scorer/picker composition for [`WorkerSelector`].
-///
-/// SelectionService constructs the concrete default state unless a caller explicitly supplies
-/// custom scorer and picker implementations through [`Self::new`].
-pub struct WorkerSelectionPolicy {
-    kv_router_config: KvRouterConfig,
-    worker_type: &'static str,
-    state: WorkerSelectionPolicyState,
-}
-
-impl WorkerSelectionPolicy {
-    pub fn new(
-        kv_router_config: KvRouterConfig,
-        worker_type: &'static str,
-        scorers: Vec<Box<dyn WorkerScorer>>,
-        picker: Box<dyn WorkerPicker>,
-    ) -> Self {
-        let picker_inputs = picker.required_worker_inputs();
-        let worker_inputs = scorers.iter().fold(picker_inputs, |inputs, scorer| {
-            inputs | scorer.required_worker_inputs()
-        });
-        Self {
-            kv_router_config,
-            worker_type,
-            state: WorkerSelectionPolicyState::Custom(RefCell::new(CustomWorkerSelectionState {
-                scorers,
-                picker,
-                worker_inputs,
-                picker_inputs,
-                candidates: Vec::new(),
-                cache_inputs: Vec::new(),
-                load_inputs: Vec::new(),
-                routing_inputs: Vec::new(),
-            })),
-        }
-    }
-
-    #[cfg_attr(not(feature = "standalone-selection"), allow(dead_code))]
-    pub(crate) fn default(kv_router_config: KvRouterConfig, worker_type: &'static str) -> Self {
-        let picker = DefaultWorkerPicker::new(kv_router_config.router_temperature);
-        Self {
-            kv_router_config,
-            worker_type,
-            state: WorkerSelectionPolicyState::Default(picker),
-        }
-    }
 }
 
 impl std::fmt::Debug for DefaultWorkerSelector {
@@ -512,7 +188,7 @@ impl DefaultWorkerScorer<KvRouterConfig> {
     }
 }
 
-fn selection_weights(
+pub(super) fn selection_weights(
     kv_router_config: &KvRouterConfig,
     request: &SchedulingRequest,
 ) -> LogitWeights {
@@ -668,238 +344,6 @@ impl DefaultWorkerPicker {
     }
 }
 
-impl<'a> WorkerSelectionInput<'a> {
-    fn new<C: WorkerConfigLike>(
-        workers: &'a HashMap<WorkerId, C>,
-        request: &'a SchedulingRequest,
-        eligibility: RoutingEligibility<'a>,
-        block_size: u32,
-        weights: LogitWeights,
-        inputs: WorkerInputs,
-    ) -> Self {
-        let min_active_prefill_tokens = if inputs.contains(WorkerInputs::LOAD)
-            && request.track_prefill_tokens
-            && weights.overlap_score_credit_decay > 0.0
-        {
-            let mut minimum = usize::MAX;
-            eligibility.for_each_eligible_worker_rank(workers, |worker, _| {
-                minimum = minimum.min(request.worker_load_for(worker).active_prefill_tokens);
-            });
-            if minimum == usize::MAX { 0 } else { minimum }
-        } else {
-            0
-        };
-        let has_tier_overlap_blocks = !request.overlap.tier_overlap_blocks.device.is_empty()
-            || !request.overlap.tier_overlap_blocks.host_pinned.is_empty()
-            || !request.overlap.tier_overlap_blocks.disk.is_empty();
-        Self {
-            request,
-            has_tier_overlap_blocks,
-            context: WorkerSelectionContext {
-                request_id: request.mode.request_id().unwrap_or("-"),
-                request_blocks: request.request_blocks(block_size),
-                block_size,
-                track_prefill_tokens: request.track_prefill_tokens,
-                weights,
-                min_active_prefill_tokens,
-                router_temperature_override: request
-                    .router_config_override
-                    .as_ref()
-                    .and_then(|config| config.router_temperature),
-            },
-        }
-    }
-
-    fn row(
-        &self,
-        worker: WorkerWithDpRank,
-        preferred_taint_multiplier: Option<f64>,
-        inputs: WorkerInputs,
-    ) -> WorkerCandidate {
-        let cached_tokens = if inputs.contains(WorkerInputs::CACHE)
-            || (inputs.contains(WorkerInputs::LOAD) && self.request.track_prefill_tokens)
-        {
-            self.request.effective_cached_tokens_for(worker)
-        } else {
-            0
-        };
-        let worker_load = if inputs.contains(WorkerInputs::LOAD) {
-            self.request.worker_loads.get(&worker).copied()
-        } else {
-            None
-        };
-        let cache = if inputs.contains(WorkerInputs::CACHE) {
-            let effective_overlap_blocks = self.request.effective_overlap_blocks_for(worker);
-            let device_overlap_blocks = self
-                .request
-                .overlap
-                .tier_overlap_blocks
-                .device
-                .get(&worker)
-                .copied()
-                .map(|blocks| blocks as f64)
-                .unwrap_or_else(|| {
-                    if self.has_tier_overlap_blocks {
-                        0.0
-                    } else {
-                        effective_overlap_blocks
-                    }
-                });
-            WorkerCacheInput {
-                effective_overlap_blocks,
-                device_overlap_blocks,
-                host_overlap_blocks: self
-                    .request
-                    .overlap
-                    .tier_overlap_blocks
-                    .host_pinned
-                    .get(&worker)
-                    .copied()
-                    .unwrap_or(0) as f64,
-                disk_overlap_blocks: self
-                    .request
-                    .overlap
-                    .tier_overlap_blocks
-                    .disk
-                    .get(&worker)
-                    .copied()
-                    .unwrap_or(0) as f64,
-                shared_beyond_device_blocks: self.request.shared_cache_hits.as_ref().map_or(
-                    0,
-                    |hits| {
-                        // `hits_beyond` expects the unweighted device prefix depth.
-                        hits.hits_beyond(device_overlap_blocks.round().max(0.0) as u32)
-                    },
-                ),
-            }
-        } else {
-            WorkerCacheInput::default()
-        };
-        let load = if inputs.contains(WorkerInputs::LOAD) {
-            let raw_prefill_tokens = if self.request.track_prefill_tokens {
-                match worker_load {
-                    Some(load) => {
-                        // Preserve the legacy operation order when overlap exceeds the prompt.
-                        let uncached_tokens = super::prefill_load::effective_prefill_tokens(
-                            self.request.isl_tokens,
-                            cached_tokens,
-                        );
-                        let projected_tokens = load.active_prefill_tokens + uncached_tokens;
-                        projected_tokens.saturating_add(cached_tokens)
-                    }
-                    None => self.request.isl_tokens,
-                }
-            } else {
-                0
-            } as f64;
-            let worker_load = worker_load.unwrap_or_default();
-            WorkerLoadInput {
-                raw_prefill_blocks: raw_prefill_tokens / self.context.block_size as f64,
-                active_prefill_tokens: worker_load.active_prefill_tokens,
-                decode_cost_blocks: worker_load.potential_decode_blocks() as f64,
-                active_requests: worker_load.active_requests,
-            }
-        } else {
-            WorkerLoadInput::default()
-        };
-
-        WorkerCandidate {
-            worker,
-            inputs,
-            cache,
-            load,
-            routing: if inputs.contains(WorkerInputs::ROUTING) {
-                WorkerRoutingInput {
-                    preferred_taint_multiplier,
-                }
-            } else {
-                WorkerRoutingInput::default()
-            },
-        }
-    }
-}
-
-fn selection_result(
-    request: &SchedulingRequest,
-    worker: WorkerWithDpRank,
-    block_size: u32,
-) -> WorkerSelectionResult {
-    WorkerSelectionResult {
-        worker,
-        required_blocks: request.request_blocks(block_size),
-        effective_overlap_blocks: request.effective_overlap_blocks_for(worker),
-        cached_tokens: request.effective_cached_tokens_for(worker),
-        potential_decode_blocks: request
-            .potential_decode_blocks_after_admission(worker, block_size),
-    }
-}
-
-fn log_selection<C: WorkerConfigLike>(
-    workers: &HashMap<WorkerId, C>,
-    request: &SchedulingRequest,
-    worker: WorkerWithDpRank,
-    worker_type: &'static str,
-    cost: f64,
-    effective_overlap_blocks: f64,
-) {
-    let request_id = request.mode.request_id().unwrap_or("-");
-    let host_pinned_blocks = request
-        .overlap
-        .tier_overlap_blocks
-        .host_pinned
-        .get(&worker)
-        .copied()
-        .unwrap_or(0);
-    let disk_blocks = request
-        .overlap
-        .tier_overlap_blocks
-        .disk
-        .get(&worker)
-        .copied()
-        .unwrap_or(0);
-
-    if request.pinned_worker == Some(worker) {
-        tracing::info!(
-            request_id,
-            "Selected pinned worker: worker_type={}, worker_id={} dp_rank={:?}, logit: {:.3}, effective cached blocks: {:.2}",
-            worker_type,
-            worker.worker_id,
-            worker.dp_rank,
-            cost,
-            effective_overlap_blocks,
-        );
-    } else if worker_type == "decode" {
-        tracing::info!(
-            router_mode = "kv",
-            request_id,
-            worker_id = worker.worker_id,
-            worker_type = %worker_type,
-            dp_rank = ?worker.dp_rank,
-            logit = cost,
-            host_pinned_blocks,
-            disk_blocks,
-            "Selected worker"
-        );
-    } else {
-        let total_kv_blocks = workers
-            .get(&worker.worker_id)
-            .and_then(WorkerConfigLike::total_kv_blocks);
-        tracing::info!(
-            router_mode = "kv",
-            request_id,
-            worker_id = worker.worker_id,
-            worker_type = %worker_type,
-            dp_rank = ?worker.dp_rank,
-            logit = cost,
-            effective_cached_blocks = effective_overlap_blocks,
-            host_pinned_blocks,
-            disk_blocks,
-            total_kv_blocks = ?total_kv_blocks,
-            "Selected worker"
-        );
-    }
-}
-
 impl<C> WorkerScorer for DefaultWorkerScorer<C>
 where
     C: Borrow<KvRouterConfig> + Send,
@@ -940,79 +384,8 @@ fn minimum_cost_index(
     best_row
 }
 
-fn collect_custom_candidates<C: WorkerConfigLike>(
-    state: &mut CustomWorkerSelectionState,
-    input: &WorkerSelectionInput<'_>,
-    workers: &HashMap<WorkerId, C>,
-    request: &SchedulingRequest,
-    eligibility: RoutingEligibility<'_>,
-) -> Result<(), KvSchedulerError> {
-    let CustomWorkerSelectionState {
-        scorers,
-        worker_inputs,
-        picker_inputs,
-        candidates,
-        cache_inputs,
-        load_inputs,
-        routing_inputs,
-        ..
-    } = state;
-    candidates.clear();
-    cache_inputs.clear();
-    load_inputs.clear();
-    routing_inputs.clear();
-    let pinned = eligibility.pinned_worker().is_some();
-    let mut error = None;
-    eligibility.any_eligible_worker_rank(workers, |worker, config| {
-        let preferred_taint_multiplier = if pinned {
-            None
-        } else {
-            request
-                .routing_constraints
-                .preferred_taint_multiplier(config.taints())
-        };
-        let candidate = input.row(worker, preferred_taint_multiplier, *worker_inputs);
-        let mut cost = 0.0;
-        for (scorer_index, scorer) in scorers.iter_mut().enumerate() {
-            let contribution = match scorer.score(&input.context, &candidate) {
-                Ok(contribution) => contribution,
-                Err(policy_error) => {
-                    error = Some(policy_error.into());
-                    return true;
-                }
-            };
-            cost += contribution;
-            if !contribution.is_finite() || !cost.is_finite() {
-                error = Some(
-                    WorkerSelectionPolicyError::NonFiniteCost {
-                        scorer_index,
-                        row: candidates.len(),
-                    }
-                    .into(),
-                );
-                return true;
-            }
-        }
-        candidates.push(ScoredWorkerCandidate { worker, cost });
-        if picker_inputs.contains(WorkerInputs::CACHE) {
-            cache_inputs.push(candidate.cache);
-        }
-        if picker_inputs.contains(WorkerInputs::LOAD) {
-            load_inputs.push(candidate.load);
-        }
-        if picker_inputs.contains(WorkerInputs::ROUTING) {
-            routing_inputs.push(candidate.routing);
-        }
-        false
-    });
-    match error {
-        Some(error) => Err(error),
-        None => Ok(()),
-    }
-}
-
 #[inline(always)]
-fn pick_default_worker<C: WorkerConfigLike>(
+pub(super) fn pick_default_worker<C: WorkerConfigLike>(
     scorer: &DefaultWorkerScorer<&KvRouterConfig>,
     picker: &DefaultWorkerPicker,
     input: &WorkerSelectionInput<'_>,
@@ -1179,35 +552,6 @@ impl WorkerPicker for DefaultWorkerPicker {
     }
 }
 
-impl<C: WorkerConfigLike> WorkerSelector<C> for WorkerSelectionPolicy {
-    #[inline(always)]
-    fn select_worker(
-        &self,
-        workers: &HashMap<WorkerId, C>,
-        request: &SchedulingRequest,
-        eligibility: RoutingEligibility<'_>,
-        block_size: u32,
-    ) -> Result<WorkerSelectionResult, KvSchedulerError> {
-        let state = match &self.state {
-            WorkerSelectionPolicyState::Default(picker) => {
-                WorkerSelectionPolicyStateRef::Default(picker)
-            }
-            WorkerSelectionPolicyState::Custom(state) => {
-                WorkerSelectionPolicyStateRef::Custom(state)
-            }
-        };
-        select_worker_with_policy(
-            &self.kv_router_config,
-            self.worker_type,
-            state,
-            workers,
-            request,
-            eligibility,
-            block_size,
-        )
-    }
-}
-
 impl<C: WorkerConfigLike> WorkerSelector<C> for DefaultWorkerSelector {
     #[inline(always)]
     fn select_worker(
@@ -1229,184 +573,16 @@ impl<C: WorkerConfigLike> WorkerSelector<C> for DefaultWorkerSelector {
     }
 }
 
-#[inline(always)]
-// DefaultWorkerSelector and SelectionService both converge here. Only the scorer/picker stage is
-// dispatched; eligibility outcomes and result construction stay host-owned and shared.
-fn select_worker_with_policy<C: WorkerConfigLike>(
-    kv_router_config: &KvRouterConfig,
-    worker_type: &'static str,
-    state: WorkerSelectionPolicyStateRef<'_>,
-    workers: &HashMap<WorkerId, C>,
-    request: &SchedulingRequest,
-    eligibility: RoutingEligibility<'_>,
-    block_size: u32,
-) -> Result<WorkerSelectionResult, KvSchedulerError> {
-    assert!(request.isl_tokens > 0);
-    eligibility.validate_pinned_worker_allowed()?;
-
-    if let Some(worker) = eligibility.pinned_worker() {
-        match eligibility.validate_worker_rank(workers, worker) {
-            Ok(_) => {}
-            Err(WorkerEligibilityError::WorkerOverloaded { .. }) => {
-                return Err(KvSchedulerError::PinnedWorkerOverloaded {
-                    worker_id: worker.worker_id,
-                });
-            }
-            Err(_) => return Err(KvSchedulerError::NoEndpoints),
-        }
-    }
-
-    let weights = selection_weights(kv_router_config, request);
-    let inputs = match &state {
-        WorkerSelectionPolicyStateRef::Default(_) => WorkerInputs::ALL,
-        WorkerSelectionPolicyStateRef::Custom(state) => RefCell::borrow(state).worker_inputs,
-    };
-    let input =
-        WorkerSelectionInput::new(workers, request, eligibility, block_size, weights, inputs);
-    let selected = match state {
-        WorkerSelectionPolicyStateRef::Default(picker) => {
-            let scorer = DefaultWorkerScorer {
-                kv_router_config,
-                worker_type,
-            };
-            pick_default_worker(&scorer, picker, &input, workers, request, eligibility)
-        }
-        WorkerSelectionPolicyStateRef::Custom(state) => {
-            let mut state = state.borrow_mut();
-            collect_custom_candidates(&mut state, &input, workers, request, eligibility)?;
-            let CustomWorkerSelectionState {
-                picker,
-                picker_inputs,
-                candidates,
-                cache_inputs,
-                load_inputs,
-                routing_inputs,
-                ..
-            } = &mut *state;
-            if candidates.is_empty() {
-                None
-            } else {
-                debug_assert!(
-                    !picker_inputs.contains(WorkerInputs::CACHE)
-                        || cache_inputs.len() == candidates.len()
-                );
-                debug_assert!(
-                    !picker_inputs.contains(WorkerInputs::LOAD)
-                        || load_inputs.len() == candidates.len()
-                );
-                debug_assert!(
-                    !picker_inputs.contains(WorkerInputs::ROUTING)
-                        || routing_inputs.len() == candidates.len()
-                );
-                let picker_input = WorkerInputView {
-                    candidates,
-                    cache: picker_inputs
-                        .contains(WorkerInputs::CACHE)
-                        .then_some(cache_inputs.as_slice()),
-                    load: picker_inputs
-                        .contains(WorkerInputs::LOAD)
-                        .then_some(load_inputs.as_slice()),
-                    routing: picker_inputs
-                        .contains(WorkerInputs::ROUTING)
-                        .then_some(routing_inputs.as_slice()),
-                };
-                let row = picker.pick(&input.context, picker_input)?;
-                let Some(candidate) = candidates.get(row) else {
-                    return Err(WorkerSelectionPolicyError::InvalidPickerRow {
-                        row,
-                        candidate_count: candidates.len(),
-                    }
-                    .into());
-                };
-                Some((candidate.worker, candidate.cost))
-            }
-        }
-    };
-    let Some((worker, cost)) = selected else {
-        if eligibility.has_eligible_worker_ignoring_overload(
-            workers
-                .iter()
-                .map(|(&worker_id, config)| (worker_id, config)),
-        ) {
-            return Err(KvSchedulerError::AllEligibleWorkersOverloaded);
-        }
-        return Err(KvSchedulerError::NoEndpoints);
-    };
-    let result = selection_result(request, worker, block_size);
-    log_selection(
-        workers,
-        request,
-        worker,
-        worker_type,
-        cost,
-        result.effective_overlap_blocks,
-    );
-    Ok(result)
-}
-
 #[cfg(test)]
 mod tests {
-    use std::collections::HashSet;
+    use std::collections::{HashMap, HashSet};
 
+    use rustc_hash::FxHashMap;
+
+    use super::super::test_support::*;
     use super::*;
-    use crate::protocols::{SharedCacheHits, WorkerConfigLike};
+    use crate::protocols::SharedCacheHits;
     use crate::scheduling::{OverlapSignals, ScheduleMode};
-
-    #[derive(Clone, Default)]
-    struct TaintedWorkerConfig {
-        taints: HashSet<String>,
-    }
-
-    impl WorkerConfigLike for TaintedWorkerConfig {
-        fn data_parallel_start_rank(&self) -> u32 {
-            0
-        }
-
-        fn data_parallel_size(&self) -> u32 {
-            1
-        }
-
-        fn max_num_batched_tokens(&self) -> Option<u64> {
-            None
-        }
-
-        fn total_kv_blocks(&self) -> Option<u64> {
-            None
-        }
-
-        fn taints(&self) -> &HashSet<String> {
-            &self.taints
-        }
-    }
-
-    fn base_request(isl_tokens: usize) -> SchedulingRequest {
-        SchedulingRequest {
-            mode: ScheduleMode::QueryOnly {
-                request_id: Some("test".into()),
-            },
-            token_seq: None,
-            isl_tokens,
-            overlap: OverlapSignals {
-                tier_overlap_blocks: Default::default(),
-                effective_overlap_blocks: HashMap::default(),
-                effective_cached_tokens: HashMap::default(),
-            },
-            worker_loads: FxHashMap::default(),
-            track_prefill_tokens: true,
-            router_config_override: None,
-            lora_name: None,
-            priority_jump: 0.0,
-            strict_priority: 0,
-            policy_class: None,
-            session_id: None,
-            expected_output_tokens: None,
-            pinned_worker: None,
-            allowed_worker_ids: None,
-            routing_constraints: crate::protocols::RoutingConstraints::default(),
-            shared_cache_hits: None,
-            resp_tx: None,
-        }
-    }
 
     fn worker_logit(
         selector: &DefaultWorkerSelector,
@@ -1430,23 +606,6 @@ mod tests {
                 &input.row(worker, None, WorkerInputs::ALL),
                 "test",
             )
-    }
-
-    fn worker_loads_with_active_decode(
-        decode_blocks: FxHashMap<WorkerWithDpRank, usize>,
-    ) -> FxHashMap<WorkerWithDpRank, crate::sequences::WorkerLoadProjection> {
-        decode_blocks
-            .into_iter()
-            .map(|(worker, active_decode_blocks)| {
-                (
-                    worker,
-                    crate::sequences::WorkerLoadProjection {
-                        active_decode_blocks,
-                        ..Default::default()
-                    },
-                )
-            })
-            .collect()
     }
 
     #[test]
@@ -2500,97 +1659,5 @@ mod tests {
             .unwrap();
 
         assert_eq!(result.worker, worker0);
-    }
-
-    #[test]
-    fn public_default_policy_matches_default_selector() {
-        let worker0 = WorkerWithDpRank::from_worker_id(0);
-        let worker1 = WorkerWithDpRank::from_worker_id(1);
-        let workers = HashMap::from([
-            (0, TaintedWorkerConfig::default()),
-            (1, TaintedWorkerConfig::default()),
-        ]);
-        let mut request = base_request(16);
-        request.worker_loads =
-            worker_loads_with_active_decode(FxHashMap::from_iter([(worker0, 8), (worker1, 1)]));
-        let config = KvRouterConfig {
-            router_temperature: 0.0,
-            ..Default::default()
-        };
-
-        let expected = DefaultWorkerSelector::new(Some(config.clone()), "test")
-            .select_worker(&workers, &request, request.eligibility(), 16)
-            .unwrap();
-        let policy = WorkerSelectionPolicy::new(
-            config.clone(),
-            "test",
-            vec![Box::new(DefaultWorkerScorer::new(config, "test"))],
-            Box::new(DefaultWorkerPicker::new(0.0)),
-        );
-        let actual = policy
-            .select_worker(&workers, &request, request.eligibility(), 16)
-            .unwrap();
-
-        assert_eq!(actual.worker, expected.worker);
-        assert_eq!(actual.required_blocks, expected.required_blocks);
-        assert_eq!(
-            actual.effective_overlap_blocks,
-            expected.effective_overlap_blocks
-        );
-        assert_eq!(actual.cached_tokens, expected.cached_tokens);
-        assert_eq!(
-            actual.potential_decode_blocks,
-            expected.potential_decode_blocks
-        );
-    }
-
-    #[test]
-    fn custom_picker_receives_requested_cache_inputs() {
-        struct HighestOverlapPicker;
-
-        impl WorkerPicker for HighestOverlapPicker {
-            fn required_worker_inputs(&self) -> WorkerInputs {
-                WorkerInputs::CACHE
-            }
-
-            fn pick(
-                &mut self,
-                _context: &WorkerSelectionContext<'_>,
-                input: WorkerInputView<'_>,
-            ) -> Result<usize, WorkerSelectionPolicyError> {
-                Ok(input
-                    .cache()
-                    .expect("requested cache inputs")
-                    .iter()
-                    .enumerate()
-                    .max_by(|(_, left), (_, right)| {
-                        left.effective_overlap_blocks()
-                            .total_cmp(&right.effective_overlap_blocks())
-                    })
-                    .map(|(row, _)| row)
-                    .expect("eligible candidate"))
-            }
-        }
-
-        let worker0 = WorkerWithDpRank::from_worker_id(0);
-        let worker1 = WorkerWithDpRank::from_worker_id(1);
-        let workers = HashMap::from([
-            (0, TaintedWorkerConfig::default()),
-            (1, TaintedWorkerConfig::default()),
-        ]);
-        let mut request = base_request(16);
-        request.overlap.effective_overlap_blocks =
-            HashMap::from([(worker0, 0.25), (worker1, 0.75)]);
-        let policy = WorkerSelectionPolicy::new(
-            KvRouterConfig::default(),
-            "test",
-            Vec::new(),
-            Box::new(HighestOverlapPicker),
-        );
-
-        let selected = policy
-            .select_worker(&workers, &request, request.eligibility(), 16)
-            .unwrap();
-        assert_eq!(selected.worker, worker1);
     }
 }
